+++ app/assets/javascripts/analytics.js
@@ -4,26 +4,32 @@ GoogleAnalytics = (function() {
   function GoogleAnalytics() {}
 
   GoogleAnalytics.load = function() {
-    var firstScript, ga;
-    window._gaq = [];
     GoogleAnalytics.analyticsId = GoogleAnalytics.getAnalyticsId();
-    window._gaq.push(["_setAccount", GoogleAnalytics.analyticsId]);
-    ga = document.createElement("script");
-    ga.type = "text/javascript";
-    ga.async = true;
-    ga.src = ("https:" === document.location.protocol ? "https://ssl" : "http://www") + ".google-analytics.com/ga.js";
-    firstScript = document.getElementsByTagName("script")[0];
-    firstScript.parentNode.insertBefore(ga, firstScript);
+    (function(i, s, o, g, r, a, m) {
+      i['GoogleAnalyticsObject'] = r;
+      i[r] = i[r] || function() {
+        (i[r].q = i[r].q || []).push(arguments);
+      };
+
+      i[r].l = 1 * new Date;
+
+      a = s.createElement(o);
+      m = s.getElementsByTagName(o)[0];
+
+      a.async = 1;
+      a.src = g;
+      m.parentNode.insertBefore(a, m);
+    })(window, document, 'script', '//www.google-analytics.com/analytics.js', 'ga');
+    ga('create', GoogleAnalytics.analyticsId, 'auto');
   };
 
   GoogleAnalytics.trackPageview = function(url) {
     if (!GoogleAnalytics.isLocalRequest()) {
-      if (url) {
-        window._gaq.push(["_trackPageview", url]);
-      } else {
-        window._gaq.push(["_trackPageview"]);
-      }
-      return window._gaq.push(["_trackPageLoadTime"]);
+      return ga('send', {
+          hitType: 'pageview',
+          page: location.pathname
+        }
+      );
     }
   };
 
@@ -50,15 +56,129 @@ Blacklight.onLoad(function(){
   }
 
   $('.iiif-dnd').on('click', function(e) {
-    e.preventDefault();
-    var link = $(e.currentTarget)
-    var manifest = link.data().manifest;
-    window._gaq.push(['_trackEvent', 'IIIF DnD', 'clicked', manifest]);
-    window.location = link.attr('href');
+    var manifest = $(e.currentTarget).data().manifest;
+    ga('send', 'event', 'IIIF DnD', 'yo', manifest, {
+      'transport': 'beacon',
+    });
   });
 
   $('.iiif-dnd').on('dragstart', function(e) {
     var manifest = $(e.currentTarget).data().manifest;
-    window._gaq.push(['_trackEvent', 'IIIF DnD', 'dragged', manifest]);
+    ga('send', 'event', 'IIIF DnD', 'dragged', manifest, {
+      'transport': 'beacon'
+    });
+  });
+
+  // Track external link clicks
+  $('a[href]:not([href^="/"],[href^="#"],[data-behavior="requests-modal"],[href=""])').on('click', function(e) {
+    ga('send', 'event', 'External link', 'clicked', this.href, {
+      'transport': 'beacon'
+    });
+    if ($('.zero-results').length > 0) {
+      ga('send', 'event', 'Zero results', 'clicked', this.href, {
+        'transport': 'beacon'
+      });
+    }
+  });
+
+  $('a[data-track="zero-results-remove-limit"]').on('click', function(e) {
+    ga('send', 'event', 'Zero results', 'clicked-remove-limit', this.href, {
+      'transport': 'beacon'
+    });
+  });
+
+  $('a[data-track="zero-results-search-all-fields"]').on('click', function(e) {
+    ga('send', 'event', 'Zero results', 'clicked-search-all-fields', this.href, {
+      'transport': 'beacon'
+    });
+  });
+
+  // When an alternate catalog is loaded, track those link clicks
+  $('[data-alternate-catalog]').on('alternateResultsLoaded', function(event) {
+    $(event.currentTarget).find('a').on('click', function(e) {
+      var eventCategory = $('.zero-results').length > 0 ? 'Zero results' : 'Alternate Results';
+      ga('send', 'event', eventCategory, 'clicked-alternate-results', this.href, {
+        'transport': 'beacon'
+      });
+    });
+  });
+
+  // Just track when a zero results page gets loaded
+  $('.zero-results').each(function() {
+    ga('send', 'event', 'Zero results', 'loaded', this.href, {
+      'transport': 'beacon'
+    });
+  });
+
+  // Featured resources on home page
+  $('.catalog-home-page .features a').on('click', function(e) {
+    ga('send', 'event', 'Featured resource', 'clicked', this.href, {
+      'transport': 'beacon'
+    });
+  });
+
+  // Side mini-nav
+  $('.side-nav-minimap button').on('click', function(e) {
+    ga('send', 'event', 'Side mini nav', $(e.currentTarget).text().trim(), location.pathname, {
+      'transport': 'beacon'
+    });
+  });
+  
+  // Facet collapse - expand events
+  $('.facet-content').on('hide.bs.collapse', function(e) {
+    ga('send', 'event', 'Facet', 'closed', $(e.currentTarget).parent().find('h3').text().trim(), {
+      'transport': 'beacon'
+    });
   })
+  $('.facet-content').on('show.bs.collapse', function(e) {
+    ga('send', 'event', 'Facet', 'expanded', $(e.currentTarget).parent().find('h3').text().trim(), {
+      'transport': 'beacon'
+    });
+  });
+
+  // Select / Select all
+  // Note: this is counted extra when select-all or unselect-all is also used
+  $('input.toggle_bookmark').on('click', function(e) {
+    ga('send', 'event', 'Selection', 'select', location.pathname, {
+      'transport': 'beacon'
+    });
+  });
+  $('span.unselect-all').on('click', function(e) {
+    ga('send', 'event', 'Selection', 'unselect-all', location.pathname, {
+      'transport': 'beacon'
+    });
+  });
+  $('span.select-all').on('click', function(e) {
+    ga('send', 'event', 'Selection', 'select-all', location.pathname, {
+      'transport': 'beacon'
+    });
+  });
+
+  // Accordion selection / collapse
+  $('button[data-accordion-section-target]').on('click', function(e) {
+    ga('send', 'event', 'Accordion selection', $(e.currentTarget).text().trim(), location.pathname, {
+      'transport': 'beacon'
+    });
+  });
+
+  // View type dropdown
+  $('#view-type-dropdown a').on('click', function(e) {
+    ga('send', 'event', 'View selection', $(e.currentTarget).text().trim(), location.pathname, {
+      'transport': 'beacon'
+    });
+  });
+
+  // Sort by dropdown
+  $('#sort-dropdown a').on('click', function(e) {
+    ga('send', 'event', 'Sort selection', $(e.currentTarget).text().trim(), location.pathname, {
+      'transport': 'beacon'
+    });
+  });
+
+  // Per page dropdown
+  $('#per_page-dropdown a').on('click', function(e) {
+    ga('send', 'event', 'Per page selection', $(e.currentTarget).text().trim(), location.pathname, {
+      'transport': 'beacon'
+    });
+  });
 });